requirements/base.txt
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
Pillow==10.4.0
drf-yasg==1.21.7
django-cors-headers==4.4.0
django-filter==24.3
argon2-cffi==23.1.0
pytz==2024.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.42
django-autoslug==1.9.9
django-countries==7.6.1
django-taggit==6.1.0
redis==5.1.0
celery==5.4.0
flower==2.0.1
django-celery-email==3.0.0
django-allauth==65.0.2
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
requests==2.32.3